backend/requirements.txt
fastapi==0.115.6
sqlmodel==0.0.22
pydantic==2.10.3
pydantic-settings==2.7.1
uvicorn==0.34.0
aiosqlite==0.20.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
better-exceptions==0.3.3
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.0
python-dotenv==1.0.1
alembic==1.14.0
anyio==4.8.0